bpf: expose is_mptcp flag to bpf_tcp_sock
is_mptcp is a field from struct tcp_sock used to indicate that the
current tcp_sock is part of the MPTCP protocol.

In this protocol, a first socket (mptcp_sock) is created with
sk_protocol set to IPPROTO_MPTCP (=262) for control purpose but it
isn't directly on the wire. This is the role of the subflow (kernel)
sockets which are classical tcp_sock with sk_protocol set to
IPPROTO_TCP. The only way to differentiate such sockets from plain TCP
sockets is the is_mptcp field from tcp_sock.

Such an exposure in BPF is thus required to be able to differentiate
plain TCP sockets from MPTCP subflow sockets in BPF_PROG_TYPE_SOCK_OPS
programs.

This patch has also been tested with CONFIG_MPTCP unset and the loading
of a BPF program accessing the is_mptcp field is failing as expected
with 'bpf verifier is misconfigured' when the verifier calls
bpf_tcp_sock_convert_ctx_access.

An example use-case is provided in
https://github.com/multipath-tcp/mptcp_net-next/tree/scripts/bpf_examples.

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@tessares.net>