mptcp: moving struct definitions
In order to access 'struct mptcp_sock' in eBPF, it is required to be
publicly accessible in the 'include/net/mptcp.h' interface rather than
being privately defined in 'net/mptcp/protocol.h'.

The other structures are required to define 'struct mptcp_sock' and the
helper is used in a newly defined BPF helper.

The moved structures are not expected to be used elsewhere than in
net/mptcp and eBPF so there is maybe a better solution to expose them to
BPF other than juste moving them here.

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@tessares.net>