bpf: Adding 'bpf_mptcp_sock' structure and helper
In order to precisely identify the parent MPTCP connection of a subflow,
it is required to access the mptcp_sock's token which uniquely identify a
MPTCP connection.

This patch adds a new structure 'bpf_mptcp_sock' exposing the 'token' field
of the 'mptcp_sock' extracted from a subflow's 'tcp_sock'. It also adds a
new BPF helper of the same name to expose the newly defined structure in
the userspace BPF API.

Signed-off-by: Nicolas Rybowski <nicolas.rybowski@tessares.net>